refactor: 입학원서 페이지의 form 부분을 컴포넌트로 분리
